requirements.txt
accelerate==0.27.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
args==0.1.0
async-timeout==5.0.1
attrs==25.3.0
braceexpand==0.1.7
certifi==2025.8.3
charset-normalizer==3.4.3
clint==0.5.1
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cmake==4.1.0
diffusers==0.26.3
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
ftfy==6.3.1
hf-xet==1.1.10
hpsv2==1.2.0
huggingface-hub==0.20.2
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
lightning-utilities==0.15.2
lit==18.1.8
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==25.0
pandas==2.3.3
Pillow==9.4.0
pluggy==1.6.0
propcache==0.3.2
protobuf==3.20.3
psutil==7.1.0
pyarrow==21.0.0
pyre-extensions==0.0.23
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
pytorch-lightning==1.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
six==1.17.0
sympy==1.14.0
timm==1.0.20
tokenizers==0.15.2
tomli==2.2.1
torch==2.0.0
torchmetrics==1.8.2
torchvision==0.15.1
tqdm==4.64.1
transformers==4.38.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
webdataset==1.0.2
xformers==0.0.17
yarl==1.20.1
zipp==3.23.0
